# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PetConnect API
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned after login or profile completion.
    SetupToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Setup token returned after OTP verification for onboarding.
  schemas:
...
    CommunityMedia:
      type: object
      properties:
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [image, video]
    CommunityPost:
      type: object
      properties:
        id:
          type: string
        author:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            username:
              type: string
            avatarUrl:
              type: string
              format: uri
        text:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/CommunityMedia"
        postType:
          type: string
          enum: [regular, share, profile_update]
        shareText:
          type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        sharesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        timeAgo:
          type: string
          example: 2 minutes ago
...
    CommunityReport:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        reportedUser:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            username:
              type: string
            avatarUrl:
              type: string
              format: uri
        count:
          type: integer
        status:
          type: string
          enum: [pending, dismissed, removed, warned]
        reasons:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CommunityPostCreateInput:
      type: object
      properties:
        text:
          type: string
          nullable: true
    CommunityPostUpdateInput:
      type: object
      properties:
        text:
          type: string
          nullable: true
...
    CommunityReportInput:
      type: object
      properties:
        reason:
          type: string
          nullable: true
...
paths:
...
  /community/posts:
    post:
      tags: [UserCommunity]
      summary: Create a community post
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostCreateInput"
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                files:
                  type: array
                  maxItems: 5
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityPost"
        "400":
          description: Validation failed
        "401":
          description: Unauthorized
    get:
      tags: [UserCommunity]
      summary: List community posts
      security:
        - AccessToken: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CommunityPost"
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
        "401":
          description: Unauthorized
...
  /community/posts/{id}:
    get:
      tags: [UserCommunity]
      summary: Get community post
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityPost"
        "404":
          description: Post not found
    patch:
      tags: [UserCommunity]
      summary: Update community post
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostUpdateInput"
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                keepMedia:
                  type: array
                  items:
                    type: string
                deleteMedia:
                  type: array
                  items:
                    type: string
                files:
                  type: array
                  maxItems: 5
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityPost"
        "404":
          description: Post not found
    delete:
      tags: [UserCommunity]
      summary: Delete community post
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post deleted
        "404":
          description: Post not found
...
  /community/posts/{id}/report:
    post:
      tags: [UserCommunity]
      summary: Report a post
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityReport"
        "400":
          description: Validation failed
  /admin/reports:
    get:
      tags: [AdminReports]
      summary: List community reports
      security:
        - AccessToken: []
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CommunityReport"
        "401":
          description: Unauthorized
  /admin/reports/{id}:
    get:
      tags: [AdminReports]
      summary: Get report details
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityReport"
        "404":
          description: Report not found
    delete:
      tags: [AdminReports]
      summary: Delete a report
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report deleted
        "404":
          description: Report not found
  /admin/reports/{id}/remove-content:
    post:
      tags: [AdminReports]
      summary: Remove reported content
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Content removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityReport"
        "404":
          description: Report not found
  /admin/reports/{id}/warn:
    post:
      tags: [AdminReports]
      summary: Warn reported user
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User warned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityReport"
        "404":
          description: Report not found
  /admin/reports/{id}/dismiss:
    post:
      tags: [AdminReports]
      summary: Dismiss a report
      security:
        - AccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report dismissed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CommunityReport"
        "404":
          description: Report not found